Group all context constants into one module
