--- utils/_context/ports.py
@@ -1,7 +1,0 @@
-from enum import IntEnum
-
-
-class ContainerPorts(IntEnum):
-    """Host ports used by tested containers"""
-
-    vcr_cassettes = 8300
